[BISERVER-2699] TimeUtils now supports more Long method parameters, fixed classes to account for the change
git-svn-id: svn://source.pentaho.org/svnroot/pentaho-open-admin-console/trunk@35937 5d404267-db13-0410-8476-b1f6b398e63a

+++ src/org/pentaho/pac/client/scheduler/ctlr/SchedulerToolbarController.java
@@ -241,7 +241,7 @@ public class SchedulerToolbarController {
       case MONTHLY: // fall through
       case YEARLY:
         if ( null == cronStr ) {
-          String repeatInterval = Integer.toString( TimeUtil.secsToMillisecs( 
+          String repeatInterval = Long.toString( TimeUtil.secsToMillisecs( 
                 scheduleEditor.getRepeatInSecs() ) );
           schedSvc.updateRepeatSchedule(
               oldSchedule.getJobName(),
@@ -529,7 +529,7 @@ public class SchedulerToolbarController {
       case MONTHLY: // fall through
       case YEARLY:
         if ( null == cronStr ) {
-          String repeatInterval = Integer.toString( TimeUtil.secsToMillisecs( 
+          String repeatInterval = Long.toString( TimeUtil.secsToMillisecs( 
                 scheduleEditor.getRepeatInSecs() ) );
           schedSvc.createRepeatSchedule(
               scheduleEditor.getName().trim(), 
@@ -621,13 +621,13 @@ public class SchedulerToolbarController {
     if ( sched.isCronSchedule() ) {
       scheduleEditor.setCronString( sched.getCronString() );  // throws CronParseException
     } else if ( sched.isRepeatSchedule() ) {
-      int repeatIntervalInSecs = TimeUtil.millsecondsToSecs( Integer.parseInt( repeatIntervalInMillisecs ) );
+      long repeatIntervalInSecs = TimeUtil.millsecondsToSecs( Long.parseLong( repeatIntervalInMillisecs ) );
       if ( 0 == repeatIntervalInSecs ) {
         // run once
         scheduleEditor.setScheduleType( ScheduleEditor.ScheduleType.RUN_ONCE );
       } else {
         // run multiple
-        scheduleEditor.setRepeatInSecs( repeatIntervalInSecs );
+        scheduleEditor.setRepeatInSecs( (int) repeatIntervalInSecs );
       }
     } else {
       throw new RuntimeException( MSGS.illegalStateMissingCronAndRepeat() );
